requirements.txt
pandas>=2.0.0
yfinance>=0.2.30
requests>=2.31.0
plotly>=5.17.0
python-dateutil>=2.8.2
pyarrow>=13.0.0
paramiko>=3.3.1
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.25.0
aiohttp>=3.9.1
tenacity>=8.2.3
tqdm>=4.66.1
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
